move: move step2 components to pages/step2 folder
