+++ src/components/pages/index.js
@@ -1,0 +1,2 @@
+export * from "./step2/KeywordButton"
+export * from "./step2/KeywordTextList"
